requirements.txt
fastapi
uvicorn
google-generativeai
inference-sdk
python-multipart
